Send email, auto-fill form, redirect, remove comment
Auto-filled the form with currently logged in name and email address.

Ability to redirect to a thank you page on successful charge.

Ability to remove a donation from the public view.

Ability to send an email confirmation with %Name% and %Amount% variables
in the outgoing message.

Ability to see the email address of the person who donated and made it a
clickable mailto link.

Added the jQuery Placeholder Plugin by Andrew Jones for the iFrame form.
Allows the Placeholder text to be visible in older browsers and IE.
http://andrew-jones.com/jquery-placeholder-plugin