Embedding the iframe and color changes
Added new shortcode to embed the iFrame. This keeps the existing design
from the iFrame popup but doesn't require an extra click to bring up the
form. Great for embedding in a "donate" page that talks about the
reasons to donate and has the form right there.

I gave default width/height for the iFrame but I also added a jQuery
plug-in to auto-resize the iFrame so it'll fit properly.

Also removed the text color 808080 from the CSS file so entered text is
easier to read.